remove asynciotest async/await workaround
