+++ src/Makefile
@@ -87,7 +87,6 @@ PY += p4p/test/test_qt.py
 PY += p4p/test/qttest.py
 
 PY += p4p/test/test_asyncio.py
-PY += p4p/test/asynciotest.py
 
 PY += p4p/example.conf
 PY += p4p/pvagw@.service
